2805: reduce memory

+++ problem/2805/main.cpp
@@ -5,9 +5,9 @@
 using namespace std;
 
 int N, M;
-vector<long long> trees;
+vector<int> trees;
 
-long long length_under(long long limit) {
+long long length_under(int limit) {
     auto up_it = upper_bound(trees.begin(), trees.end(), limit);
     long long sum = 0;
     for (auto it = up_it; it != trees.end(); it++) {
@@ -20,16 +20,16 @@ int main() {
     ios_base::sync_with_stdio(false);
     cin.tie(nullptr);
     cin >> N >> M;
-    long long num;
+    int num;
     for (int i = 0; i < N; i++) {
         cin >> num;
         trees.push_back(num);
     }
     sort(trees.begin(), trees.end());
-    long long lo = 0;
-    long long hi = *prev(trees.end());
+    int lo = 0;
+    int hi = *prev(trees.end());
     while (lo <= hi) {
-        long long mid = (lo + hi) / 2;
+        int mid = (lo + hi) / 2;
         if (length_under(mid) < M) {
             hi = mid - 1;
         } else {
